LayoutUnit(unsigned long long value) compile failure https://bugs.webkit.org/show_bug.cgi?id=106239
Cast the result of ('unsigned long long' * 'int') to an int before storing in m_value
to avoid compile failure on some platforms.

Reviewed by Ryosuke Niwa.

* platform/LayoutUnit.h:
(WebCore::LayoutUnit::LayoutUnit): Cast to int after multiplication.


Canonical link: https://commits.webkit.org/124440@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@138978 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/LayoutUnit.h
@@ -86,7 +86,7 @@ public:
         m_value = clampTo<int>(value * kEffectiveFixedPointDenominator);
 #else
         REPORT_OVERFLOW(isInBounds(static_cast<unsigned>(value)));
-        m_value = value * kEffectiveFixedPointDenominator;
+        m_value = static_cast<int>(value * kEffectiveFixedPointDenominator);
 #endif
     }
     LayoutUnit(float value)
